move compare.h and symbols.h into private
