+++ src/parse.cc
@@ -10,11 +10,11 @@
 #include <iostream>
 #include <cctype>
 
-#include "internal/symbols.h"
 #include "parse.h"
 #include "parser/identifier.h"
 #include "parser/keyword.h"
 #include "parser/statements/create_database.h"
+#include "parser/symbol.h"
 
 namespace sqltoast {
